requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
alembic
python-dotenv
psycopg2-binary
python-jose
passlib[bcrypt] 